Refactor InventoryConfig subclasses into separate classes for readability Accordions/Subcategories
